+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 - Drop deprecated `.to_blob()` & `to_canvas()` (use `.toBlob()` & `toCanvas()`
   instead).
-
+- Added separate `._calculate_size()` to simplify logic override.
 
 1.0.7 / 2020-08-20
 ------------------